refactor: porra toda

+++ db/migrate/20250812184406_create_g_sexos.rb
@@ -1,18 +1,19 @@
 # frozen_string_literal: true
 
 class CreateGSexos < ActiveRecord::Migration[7.2]
-  def up
-    create_table :g_sexos do |t|
-      t.string :descricao
-          
-      t.string :created_by
-      t.string :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+  unless table_exists?(:g_sexos)
+    def up
+      create_table :g_sexos do |t|
+        t.string :descricao
+        t.string :created_by
+        t.string :updated_by
+        t.datetime :deleted_at
+        t.timestamps
+      end
     end
   end
 
   def down
-    drop_table :g_sexos
+    drop_table :g_sexos if table_exists?(:g_sexos)
   end
 end
